+++ app/Http/Controllers/Admin/AdminQuestionController.php
@@ -239,19 +239,13 @@ class AdminQuestionController extends AdminController
     /**
      * Show a list of all the questions formatted for Datatables.
      *
-     * @param \Illuminate\Http\Request     $request
      * @param \Yajra\Datatables\Datatables $dataTable
      *
      * @return \Illuminate\Contracts\Routing\ResponseFactory|\Symfony\Component\HttpFoundation\Response
      * @throws \Exception
      */
-    public function data(Request $request, Datatables $dataTable)
+    public function data(Datatables $dataTable)
     {
-        // Disable this query if isn't AJAX
-        if (! $request->ajax()) {
-            return response('Forbidden.', 403);
-        }
-
         $question = Question::select([
             'id',
             'short_name',
@@ -265,7 +259,7 @@ class AdminQuestionController extends AdminController
             'unpublish' => '<span class="label label-warning">'.strval(__('admin/question/model.status_list.unpublish')).'</span>',
         ];
 
-        return $dataTable->of($question)
+        return $dataTable->eloquent($question)
             ->editColumn('status', function (Question $question) use ($statusLabel) {
                 return $statusLabel[$question->status];
             })
@@ -277,6 +271,6 @@ class AdminQuestionController extends AdminController
             })
             ->rawColumns(['actions', 'status'])
             ->removeColumn('id')
-            ->make(true);
+            ->toJson();
     }
 }