+++ resources/lang/en/admin/level/title.php
@@ -1,9 +1,10 @@
 <?php
 
 return [
-    'level_management'   => 'Level Management',
-    'level_show'         => 'Level Details',
+    'level_management' => 'Level Management',
+    'level_show' => 'Level Details',
     'create_a_new_level' => 'Create a New Level',
-    'level_update'       => 'Update a Level',
-    'level_delete'       => 'Delete a Level',
+    'create_a_new_level_desc' => 'create a new level',
+    'level_update' => 'Update a Level',
+    'level_delete' => 'Delete a Level',
 ];