+++ resources/lang/en/admin/user/title.php
@@ -1,9 +1,10 @@
 <?php
 
 return [
-    'user_management'   => 'User Management',
-    'user_show'         => 'User Details',
+    'user_management' => 'User Management',
+    'user_management_desc' => 'create and edit users',
+    'user_show' => 'User Details',
     'create_a_new_user' => 'Create a New User',
-    'user_update'       => 'Update User',
-    'user_delete'       => 'Delete User',
+    'user_update' => 'Update User',
+    'user_delete' => 'Delete User',
 ];