+++ resources/views/admin/level/create.blade.php
@@ -1,14 +1,12 @@
 @extends('layouts.admin')
 
 {{-- Web site Title --}}
-@section('title')
-    @lang('admin/level/title.create_a_new_level') :: @parent
-@endsection
+@section('title', __('admin/level/title.create_a_new_level'))
 
 {{-- Content Header --}}
 @section('header')
     @lang('admin/level/title.create_a_new_level')
-    <small>create a new level</small>
+    <small>@lang('admin/level/title.create_a_new_level_desc')</small>
 @endsection
 
 {{-- Breadcrumbs --}}
@@ -26,15 +24,117 @@
     <li class="active">
         @lang('admin/level/title.create_a_new_level')
     </li>
-    @endsection
+@endsection
 
-    {{-- Content --}}
-    @section('content')
+{{-- Content --}}
+@section('content')
 
-            <!-- Notifications -->
+    <!-- Notifications -->
     @include('partials.notifications')
-            <!-- ./ notifications -->
+    <!-- ./ notifications -->
 
-    @include('admin/level/_form')
+    {!! Form::open(['route' => ['admin.levels.store'], 'method' => 'post', 'files' => true]) !!}
+
+    <div class="box box-solid">
+        <div class="box-body">
+
+            <div class="row">
+                <div class="col-xs-6">
+
+                    <!-- name -->
+                    <div class="form-group {{ $errors->has('name') ? 'has-error' : '' }}">
+                        {!! Form::label('name', __('admin/level/model.name'), ['class' => 'control-label required']) !!}
+                        <div class="controls">
+                            {!! Form::text('name', null, ['class' => 'form-control', 'required' => 'required']) !!}
+                            <span class="help-block">{{ $errors->first('name', ':message') }}</span>
+                        </div>
+                    </div>
+                    <!-- ./ name -->
+
+                    <!-- required_points -->
+                    <div class="form-group {{ $errors->has('required_points') ? 'has-error' : '' }}">
+                        {!! Form::label('required_points', __('admin/level/model.required_points'), ['class' => 'control-label required']) !!}
+                        <div class="controls">
+                            {!! Form::number('required_points', null, ['class' => 'form-control', 'required' => 'required', 'min' => '1']) !!}
+                            <span class="help-block">{{ $errors->first('required_points', ':message') }}</span>
+                        </div>
+                    </div>
+                    <!-- ./ required_points -->
+
+                    <!-- activation status -->
+                    <div class="form-group {{ $errors->has('active') ? 'has-error' : '' }}">
+                        {!! Form::label('active', __('admin/level/model.active'), ['class' => 'control-label required']) !!}
+                        <div class="controls">
+                            {!! Form::select('active', ['1' => __('general.yes'), '0' => __('general.no')], null, ['class' => 'form-control', 'required' => 'required']) !!}
+                            {{ $errors->first('active', '<span class="help-inline">:message</span>') }}
+                        </div>
+                    </div>
+                    <!-- ./ activation status -->
+
+                </div>
+                <div class="col-xs-6">
+
+                    <!-- image -->
+                {{--
+                <div class="form-group {{ $errors->has('image') ? 'has-error' : '' }}">
+                    {!! Form::label('image', __('admin/level/model.image'), ['class' => 'control-label required']) !!}
+                    <p class="text-muted">@lang('admin/level/model.image_help')</p>
+                    <div class="controls">
+                        <div class="fileinput fileinput-new" data-provides="fileinput">
+                            <div class="fileinput-preview thumbnail" data-trigger="fileinput"
+                                 style="width: 150px; height: 150px;">
+                                @if (isset($level))
+                                    <img src="{{ $level->image->url() }}">
+                                @endif
+                            </div>
+                            <p>
+                            <span class="btn btn-default btn-file">
+                                <span class="fileinput-new"><i
+                                            class="fa fa-picture-o"></i> @lang('button.pick_image')</span>
+                                <span class="fileinput-exists"><i
+                                            class="fa fa-picture-o"></i> @lang('button.upload_image')</span>
+                                {!! Form::file('image') !!}
+                            </span>
+                                <a href="#" class="btn fileinput-exists btn-default" data-dismiss="fileinput">
+                                    <i class="fa fa-times"></i> @lang('button.delete_image')
+                                </a>
+                            </p>
+                        </div>
+                    </div>
+                    <span class="help-block">{{ $errors->first('image', ':message') }}</span>
+                </div>
+            </div>
+            --}}
+                <!-- ./ image -->
+
+                </div>
+            </div>
+
+            <div class="box-footer">
+                <!-- Form Actions -->
+                <a href="{{ route('admin.levels.index') }}">
+                    <button type="button" class="btn btn-primary">
+                        <i class="fa fa-arrow-left"></i> @lang('general.back')
+                    </button>
+                </a>
+            {!! Form::button(__('button.save'), ['type' => 'submit', 'class' => 'btn btn-success']) !!}
+            <!-- ./ form actions -->
+            </div>
+        </div>
+    </div>
+    {!! Form::close() !!}
+
+    {{-- Styles --}}
+    @push('styles')
+        <!-- File Input -->
+        {{-- <link rel="stylesheet" type="text/css" href="{{ asset('vendor/jasny-bootstrap/dist/css/jasny-bootstrap.min.css') }}"> --}}
+    @endpush
+
+    {{-- Scripts --}}
+    @push('scripts')
+        <!-- File Input -->
+        {{-- <script type="text/javascript" src="{{ asset('vendor/jasny-bootstrap/dist/js/jasny-bootstrap.min.js') }}"></script> --}}
+    @endpush
+
 
 @endsection